server/requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
elasticsearch-dsl==7.3.0
psycopg2-binary==2.8.6
django-debug-toolbar==3.2